fix tests and added in template Auth
