+++ frontend/package.json
@@ -38,6 +38,7 @@
         "@testing-library/dom": "^8.1.0",
         "@testing-library/jest-dom": "5.9.0",
         "@testing-library/react": "10.2.1",
+        "@testing-library/user-event": "^13.1.9",
         "@types/jest": "^26.0.24",
         "@types/react": "^17.0.14",
         "@types/react-dom": "^17.0.9",
