+++ frontend/src/Components/Molecules/AuthForm/AuthForm.tsx
@@ -1,0 +1,11 @@
+import { memo, ReactNode } from 'react'
+
+import { StyledForm } from './AuthForm.styles'
+
+interface IFormProps {
+    children: ReactNode
+}
+
+export const AuthForm = memo(({ children }: IFormProps) => {
+    return <StyledForm>{children}</StyledForm>
+})
