--- frontend/src/Components/Molecules/Form/Form.tsx
@@ -1,11 +1,0 @@
-import { memo, ReactNode } from "react"
-
-import { StyledForm } from "./Form.styles";
-
-interface IFormProps {
-    children: ReactNode
-}
-
-export const Form = memo(({ children }: IFormProps)  => {
-    return <StyledForm>{children}</StyledForm>
-});
