+++ frontend/src/Components/Templates/Auth/Auth.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Auth = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Auth
