+++ frontend/src/Pages/authPage/index.tsx
@@ -1,0 +1,11 @@
+// import 
+
+const Auth = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Auth
